LayoutTests/fast/css/quirks-mode-ignore-list-item.html


<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<style>
div {
  display:list-item;
  float:right
}
</style>
<div id="target"></div>
<script>
test(function() {
  assert_equals(getComputedStyle(target).display, "list-item");
  assert_equals(getComputedStyle(target).float, "right");
}, "Ensure that elements with display:list-item retain their display value and are not set to 'display:block' when float is set in quirks mode.");
</script>
